pyproject.toml
[project]
name = "openrouter-poc-01"
version = "0.1.0"
description = "OpenRouter experiments"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openrouter",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main"]
